+++ aleph/logic/sessions.py
@@ -12,7 +12,7 @@ def get_jwt_secret():
 
 
 def create_token(role):
-    from aleph.views.serializers import RoleSchema
+    from aleph.serializers import RoleSchema
     role, _ = RoleSchema().dump(role)
     role.pop('foreign_id', None)
     role.pop('created_at', None)
