+++ aleph/serializers/__init__.py
@@ -1,72 +1,17 @@
 from flask import request
-from followthemoney import model
-from exactitude import countries, languages, dates
 from marshmallow import Schema, post_dump
 from marshmallow.fields import Nested, Integer, String, DateTime, List
-from marshmallow.fields import Raw, Dict, Boolean, Float
-from marshmallow.exceptions import ValidationError
+from marshmallow.fields import Dict, Boolean, Float
 from marshmallow.validate import Email, Length
 
-from aleph.core import url_for, get_config
+from aleph.core import url_for
 from aleph.logic.collections import collection_url
 from aleph.logic.entities import entity_url
 from aleph.logic.documents import document_url
+from aleph.serializers.common import BaseSchema, SchemaName, PartialDate
+from aleph.serializers.common import Category, Country, Language
 from aleph.model import Role, Document, Entity, Collection
 from aleph.util import ensure_list
-
-
-class Category(String):
-    """A category of collections, e.g. leaks, court cases, sanctions list."""
-
-    def _validate(self, value):
-        categories = get_config('COLLECTION_CATEGORIES', {})
-        if value not in categories.keys():
-            raise ValidationError('Invalid category.')
-
-
-class Language(String):
-    """A valid language code."""
-
-    def _validate(self, value):
-        if not languages.validate(value):
-            raise ValidationError('Invalid language code.')
-
-
-class Country(String):
-    """A valid country code."""
-
-    def _validate(self, value):
-        if not countries.validate(value):
-            raise ValidationError('Invalid country code: %s' % value)
-
-
-class PartialDate(String):
-    """Any valid prefix of an ISO 8601 datetime string."""
-
-    def _validate(self, value):
-        if not dates.validate(value):
-            raise ValidationError('Invalid date: %s' % value)
-
-
-class SchemaName(String):
-
-    def _validate(self, value):
-        schema = model.get(value)
-        if schema is None:
-            raise ValidationError('Invalid schema name: %s' % value)
-
-
-class BaseSchema(Schema):
-    EXPAND = []
-
-    id = String(dump_only=True)
-    score = Float(dump_only=True)
-    highlight = String(dump_only=True)
-
-    # these are raw because dumping fails if the dates are already strings, as
-    # in the case of data coming from the ES index.
-    created_at = Raw(dump_only=True)
-    updated_at = Raw(dump_only=True)
 
 
 class RoleSchema(BaseSchema):
